requirements.txt
Django
psycopg2
gunicorn
South
django-tastypie
django-mptt
requests